+++ NewCodeCell.ps1
@@ -1,19 +1,35 @@
 function New-CodeCell {
     param(
         [Parameter(Mandatory)]
-        $Source
-    )    
-    @"
-{
+        $Source,
+        [Switch]$DotNetInteractive
+    )
+
+    $DotNetInteractiveMetadata = ''
+    if ($DotNetInteractive) {
+        $DotNetInteractiveMetadata = @'
+        "dotnet_interactive": {
+            "language": "pwsh"
+          },
+'@        
+    }
+
+    $targetSource = @($source.split("`n")) | ConvertTo-Json
+
+    $result = @"
+{{
     "cell_type": "code",
     "execution_count": 0,
-    "metadata": {
-     "tags": [
-      "injected-parameters"
-     ]
-    },
+    "metadata": {{
+        {0}
+        "tags": [
+        "injected-parameters"
+        ]
+    }},
     "outputs": [],
-    "source": $(@($source.split("`n")) | ConvertTo-Json)    
-}
-"@
+    "source": {1}
+}}    
+"@ -f $DotNetInteractiveMetadata, $targetSource
+
+    $result
 }